refactor: constants 폴더명 constant로 변경
