+++ src/components/PrivateLayout.tsx
@@ -6,8 +6,6 @@ export default function PrivateLayout() {
   const { data: userData } = useAppSelector(state => state.userReducer);
 
   if (Object.keys(userData).length === 0) {
-    console.log('not user');
-
     return <Navigate to={PATH.signIn} replace />;
   }
