+++ src/pages/EditPasswordPage.tsx
@@ -11,7 +11,7 @@ import SignInput from 'components/@common/SignInput';
 import { PATH } from 'Router';
 import { useDispatch } from 'react-redux';
 import { updateSnackBar } from 'redux/actions/snackBar';
-import { MESSAGE } from 'constants/message';
+import { MESSAGE } from 'constant/message';
 
 const EditPasswordPage = () => {
   const navigate = useNavigate();
